Moved project to Visual Studio 2019 OpenMP needed to be disabled in order to do this.

+++ SphericalHarmonicsLighting/HDRProbeLight.cpp
@@ -27,8 +27,8 @@ glm::vec3 HDRProbeLight::get_light(float theta, float phi) const {
 	const float u = direction.x * r;
 	const float v = direction.y * r;
 
-	const int x = (u * 0.5f + 0.5f) * size;
-	const int y = (v * 0.5f + 0.5f) * size;
+	const int x = (int)(u * 0.5f + 0.5f) * size;
+	const int y = (int)(v * 0.5f + 0.5f) * size;
 
 	assert(data);
 	return data[x * size + y];
